feat: move TodoList modal implementation to new TodoList page component (refs #79)
